Add api and cli for module tagging
