+++ packages/react-server-module-tagger/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "react-server-module-tagger",
+  "version": "0.0.1",
+  "description": "calculates module level config",
+  "main": "index.js",
+  "scripts": {
+    "test": "ava test.js",
+    "clean": "del index.js npm-debug.log*"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+ssh://git@github.com/redfin/react-server.git"
+  },
+  "keywords": [
+    "react-server"
+  ],
+  "author": "Doug Wade <doug.wade@redfin.com>",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/redfin/react-server/issues"
+  },
+  "homepage": "https://github.com/redfin/react-server#readme",
+  "devDependencies": {
+    "ava": "^0.15.2",
+    "babel-cli": "^6.9.0",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "del-cli": "^0.2.0"
+  }
+}
